feat: end-to-end scenario-based AI recommendations using live DB data

+++ backend/src/main/java/com/codecool/clutchmind/model/PossessionEvent.java
@@ -12,7 +12,7 @@ public record PossessionEvent(
         int fg3a,
         int fg3m,
         String gamedate,
-        int gameid,
+        long gameid,
         @JsonProperty("nonshootingfoulsthatresultedinfts")
         int nonShootingFoulsThatResultedInFts,
         int offensiverebounds,
